+++ HISTORY.md
@@ -41,7 +41,8 @@
 ### 2021-10-07
 * objects, byob: new BlockVisibilityDialogMorph for bulk-selecting blocks to hide / show in the palette
 * ojects: simplified palette context menu
-* blocks: removed "hide" option from context menu of primitive blocks in the palette 
+* blocks: removed "hide" option from context menu of primitive blocks in the palette
+* objects, byob: optimized bulk hiding & showing palette blocks 
 
 ### 2021-10-06
 * threads: programmatically hide individual variables in palette